Split out per-operation implementations into standalone files
This has been done for the last few added operations and keeps the
`CurlOps.cc` size modest.
