+++ src/XrdClCurl/CurlOps.cc
@@ -20,22 +20,14 @@
 #include "../common/CurlResponses.hh"
 #include "CurlUtil.hh"
 #include "CurlWorker.hh"
-#include "OptionsCache.hh"
-#include "CurlFile.hh"
 
 #include <XrdCl/XrdClDefaultEnv.hh>
 #include <XrdCl/XrdClLog.hh>
 #include <XrdCl/XrdClXRootDResponses.hh>
 
-#include <curl/curl.h>
-#include <tinyxml2.h>
-
 #include <unistd.h>
 #include <cmath>
 #include <utility>
-#include <sstream>
-#include <stdexcept>
-#include <vector>
 
 using namespace XrdClCurl;
 
@@ -359,555 +351,4 @@ CurlOperation::WaitSocketCallback(std::string &err)
         m_logger->Debug(kLogXrdClCurl, "Got reverse connection on socket %d", m_broker_reverse_socket);
     }
     return m_broker_reverse_socket;
-}
-
-// OPTIONS information is available.
-//
-// Reconfigure curl handle, as necessary, to use PROPFIND
-void
-CurlStatOp::OptionsDone()
-{
-    auto &instance = VerbsCache::Instance();
-    auto verbs = instance.Get(m_url);
-    if (verbs.IsSet(VerbsCache::HttpVerb::kPROPFIND)) {
-        curl_easy_setopt(m_curl.get(), CURLOPT_CUSTOMREQUEST, "PROPFIND");
-        curl_easy_setopt(m_curl.get(), CURLOPT_NOBODY, 0L);
-        m_is_propfind = true;
-    } else {
-        m_is_propfind = false;
-        curl_easy_setopt(m_curl.get(), CURLOPT_NOBODY, 1L);
-    }
-}
-
-CurlOperation::RedirectAction
-CurlStatOp::Redirect(std::string &target)
-{
-    auto headers = m_headers;
-    auto result = CurlOperation::Redirect(target);
-    if (result == CurlOperation::RedirectAction::Fail) {
-        return result;
-    }
-    auto &instance = VerbsCache::Instance();
-    auto verbs = instance.Get(target);
-    if (verbs.IsSet(VerbsCache::HttpVerb::kUnset)) {
-        m_headers = std::move(headers);
-        return CurlOperation::RedirectAction::ReinvokeAfterAllow;
-    }
-
-    if (verbs.IsSet(VerbsCache::HttpVerb::kPROPFIND)) {
-        curl_easy_setopt(m_curl.get(), CURLOPT_CUSTOMREQUEST, "PROPFIND");
-        curl_easy_setopt(m_curl.get(), CURLOPT_NOBODY, 0L);
-        m_is_propfind = true;
-    } else {
-        m_is_propfind = false;
-        curl_easy_setopt(m_curl.get(), CURLOPT_NOBODY, 1L);
-    }
-    return CurlOperation::RedirectAction::Reinvoke;
-}
-
-void
-CurlStatOp::Setup(CURL *curl, CurlWorker &worker)
-{
-    CurlOperation::Setup(curl, worker);
-    curl_easy_setopt(m_curl.get(), CURLOPT_WRITEFUNCTION, CurlStatOp::WriteCallback);
-    curl_easy_setopt(m_curl.get(), CURLOPT_WRITEDATA, this);
-
-    auto &instance = VerbsCache::Instance();
-    auto verbs = instance.Get(m_url);
-    if (verbs.IsSet(VerbsCache::HttpVerb::kPROPFIND)) {
-        curl_easy_setopt(m_curl.get(), CURLOPT_CUSTOMREQUEST, "PROPFIND");
-        curl_easy_setopt(m_curl.get(), CURLOPT_NOBODY, 0L);
-        m_is_propfind = true;
-    } else {
-        curl_easy_setopt(m_curl.get(), CURLOPT_NOBODY, 1L);
-    }
-}
-
-void
-CurlStatOp::ReleaseHandle()
-{
-    if (m_curl == nullptr) return;
-    curl_easy_setopt(m_curl.get(), CURLOPT_NOBODY, 0L);
-    if (m_is_propfind) {
-        curl_easy_setopt(m_curl.get(), CURLOPT_CUSTOMREQUEST, nullptr);
-    }
-    curl_easy_setopt(m_curl.get(), CURLOPT_WRITEFUNCTION, nullptr);
-    curl_easy_setopt(m_curl.get(), CURLOPT_WRITEDATA, nullptr);
-    CurlOperation::ReleaseHandle();
-}
-
-size_t
-CurlStatOp::WriteCallback(char *buffer, size_t size, size_t nitems, void *this_ptr)
-{
-    auto me = static_cast<CurlStatOp*>(this_ptr);
-    if (me->m_is_propfind) {
-        if (size * nitems + me->m_response.size() > 1'000'000) {
-            me->m_logger->Error(kLogXrdClCurl, "Response too large for PROPFIND operation");
-            return 0;
-        }
-        me->m_response.append(buffer, size * nitems);
-    }
-    return size * nitems;
-}
-
-std::pair<int64_t, bool>
-CurlStatOp::ParseProp(tinyxml2::XMLElement *prop) {
-    if (prop == nullptr) {
-        return {-1, false};
-    }
-    for (auto child = prop->FirstChildElement(); child != nullptr; child = child->NextSiblingElement()) {
-        if (!strcmp(child->Name(), "D:getcontentlength") || !strcmp(child->Name(), "lp1:getcontentlength")) {
-            auto len = child->GetText();
-            if (len) {
-                m_length = std::stoll(len);
-            }
-        } else if (!strcmp(child->Name(), "D:resourcetype") || !strcmp(child->Name(), "lp1:resourcetype")) {
-            m_is_dir = child->FirstChildElement("D:collection") != nullptr;
-        }
-    }
-    return {m_length, m_is_dir};
-}
-
-std::pair<int64_t, bool>
-CurlStatOp::GetStatInfo() {
-    if (!m_is_propfind) {
-        m_length = m_headers.GetContentLength();
-        return {m_length, false};
-    }
-    if (m_length >= 0) {
-        return {m_length, m_is_dir};
-    }
-
-    tinyxml2::XMLDocument doc;
-    auto err = doc.Parse(m_response.c_str());
-    if (err != tinyxml2::XML_SUCCESS) {
-        m_logger->Error(kLogXrdClCurl, "Failed to parse XML response: %s", m_response.substr(0, 1024).c_str());
-        return {-1, false};
-    }
-
-    auto elem = doc.RootElement();
-    if (strcmp(elem->Name(), "D:multistatus")) {
-        m_logger->Error(kLogXrdClCurl, "Unexpected XML response: %s", m_response.substr(0, 1024).c_str());
-        return {-1, false};
-    }
-    auto found_response = false;
-    for (auto response = elem->FirstChildElement(); response != nullptr; response = response->NextSiblingElement()) {
-        if (!strcmp(response->Name(), "D:response")) {
-            found_response = true;
-            elem = response;
-            break;
-        }
-    }
-    if (!found_response) {
-        m_logger->Error(kLogXrdClCurl, "Failed to find response element in XML response: %s", m_response.substr(0, 1024).c_str());
-        return {-1, false};
-    }
-    for (auto child = elem->FirstChildElement(); child != nullptr; child = child->NextSiblingElement()) {
-		if (strcmp(child->Name(), "D:propstat")) {
-            continue;
-        }
-        for (auto prop = child->FirstChildElement(); prop != nullptr; prop = prop->NextSiblingElement()) {
-            if (!strcmp(prop->Name(), "D:prop")) {
-                return ParseProp(prop);
-            }
-        }
-	}
-    m_logger->Error(kLogXrdClCurl, "Failed to find properties in XML response: %s", m_response.substr(0, 1024).c_str());
-    return {-1, false};
-}
-
-bool CurlStatOp::RequiresOptions() const
-{
-    auto &instance = VerbsCache::Instance();
-    auto verbs = instance.Get(m_url);
-    return verbs.IsSet(VerbsCache::HttpVerb::kUnset);
-}
-
-void CurlStatOp::Success()
-{
-    SuccessImpl(true);
-}
-
-void
-CurlStatOp::SuccessImpl(bool returnObj)
-{
-    SetDone(false);
-    m_logger->Debug(kLogXrdClCurl, "CurlStatOp::Success");
-    if (m_handler == nullptr) {return;}
-    XrdCl::AnyObject *obj = nullptr;
-    if (returnObj) {
-        auto [size, isdir] = GetStatInfo();
-        if (size < 0) {
-            m_logger->Error(kLogXrdClCurl, "Failed to get stat info for %s", m_url.c_str());
-            Fail(XrdCl::errErrorResponse, kXR_FSError, "Server responded without object size");
-            return;
-        }
-        if (m_is_propfind) {
-            m_logger->Debug(kLogXrdClCurl, "Successful propfind operation on %s (size %lld, isdir %d)", m_url.c_str(), static_cast<long long>(size), isdir);
-        } else {
-            m_logger->Debug(kLogXrdClCurl, "Successful stat operation on %s (size %lld)", m_url.c_str(), static_cast<long long>(size));
-        }
-
-        XrdCl::StatInfo *stat_info;
-        if (m_response_info){
-            auto info = new XrdClCurl::StatResponse("nobody", size,
-            XrdCl::StatInfo::Flags::IsReadable | (isdir ? XrdCl::StatInfo::Flags::IsDir : 0), time(NULL));
-            info->SetResponseInfo(MoveResponseInfo());
-            stat_info = info;
-        } else {
-            stat_info = new XrdCl::StatInfo("nobody", size,
-            XrdCl::StatInfo::Flags::IsReadable | (isdir ? XrdCl::StatInfo::Flags::IsDir : 0), time(NULL));
-        }
-        obj = new XrdCl::AnyObject();
-        obj->Set(stat_info);
-    } else if (m_response_info) {
-        auto info = new XrdClCurl::OpenResponseInfo();
-        info->SetResponseInfo(MoveResponseInfo());
-        obj = new XrdCl::AnyObject();
-    }
-
-    auto handle = m_handler;
-    m_handler = nullptr;
-    handle->HandleResponse(new XrdCl::XRootDStatus(), obj);
-}
-
-CurlOpenOp::CurlOpenOp(XrdCl::ResponseHandler *handler, const std::string &url, struct timespec timeout,
-    XrdCl::Log *logger, XrdClCurl::File *file, bool response_info)
-:
-    CurlStatOp(handler, url, timeout, logger, response_info),
-    m_file(file)
-{}
-
-void
-CurlOpenOp::ReleaseHandle()
-{
-    curl_easy_setopt(m_curl.get(), CURLOPT_OPENSOCKETFUNCTION, nullptr);
-    curl_easy_setopt(m_curl.get(), CURLOPT_OPENSOCKETDATA, nullptr);
-    curl_easy_setopt(m_curl.get(), CURLOPT_SOCKOPTFUNCTION, nullptr);
-    curl_easy_setopt(m_curl.get(), CURLOPT_SOCKOPTDATA, nullptr);
-    CurlStatOp::ReleaseHandle();
-}
-
-void
-CurlOpenOp::SetOpenProperties()
-{
-    char *url = nullptr;
-    curl_easy_getinfo(m_curl.get(), CURLINFO_EFFECTIVE_URL, &url);
-    if (url && m_file) {
-        m_file->SetProperty("LastURL", url);
-    }
-    if (UseX509Auth() && m_file) {
-        m_file->SetProperty("UseX509Auth", "true");
-    }
-    const auto &broker = GetBrokerUrl();
-    if (!broker.empty() && m_file) {
-        m_file->SetProperty("BrokerURL", broker);
-    }
-}
-
-void
-CurlOpenOp::Success()
-{
-    SetDone(false);
-    SetOpenProperties();
-    auto [size, isdir] = GetStatInfo();
-    if (isdir) {
-        m_logger->Error(kLogXrdClCurl, "Cannot open a directory");
-        Fail(XrdCl::errErrorResponse, kXR_isDirectory, "Cannot open a directory");
-        return;
-    }
-    if (size >= 0) {
-        m_file->SetProperty("ContentLength", std::to_string(size));
-    }
-    SuccessImpl(false);
-}
-
-void
-CurlOpenOp::Fail(uint16_t errCode, uint32_t errNum, const std::string &msg)
-{
-    // Note: OpenFlags::New is equivalent to O_CREAT | O_EXCL; OpenFlags::Write is equivalent to O_WRONLY | O_CREAT;
-    // OpenFlags::Delete is equivalent to O_CREAT | O_TRUNC;
-    if (errCode == XrdCl::errErrorResponse &&  errNum == kXR_NotFound && (m_file->Flags() & (XrdCl::OpenFlags::New | XrdCl::OpenFlags::Write | XrdCl::OpenFlags::Delete))) {
-        m_logger->Debug(kLogXrdClCurl, "CurlOpenOp succeeds as 404 was expected");
-        SetOpenProperties();
-        SuccessImpl(false);
-        m_file->SetProperty("ContentLength", "0");
-        return;
-    }
-    CurlOperation::Fail(errCode, errNum, msg);
-}
-
-CurlCopyOp::CurlCopyOp(XrdCl::ResponseHandler *handler, const std::string &source_url, const Headers &source_hdrs,
-const std::string &dest_url, const Headers &dest_hdrs, struct timespec timeout, XrdCl::Log *logger) :
-    CurlOperation(handler, dest_url, timeout, logger),
-    m_source_url(source_url),
-    m_header_list(nullptr, &curl_slist_free_all)
-{
-    m_minimum_rate = 1;
-
-    for (const auto &info : source_hdrs) {
-        m_header_list.reset(curl_slist_append(m_header_list.release(),
-            (std::string("TransferHeader") + info.first + ": " + info.second).c_str()
-        ));
-    }
-    for (const auto &info : dest_hdrs) {
-        m_header_list.reset(curl_slist_append(m_header_list.release(),
-            (info.first + ": " + info.second).c_str()
-        ));
-    }
-}
-
-void
-CurlCopyOp::Setup(CURL *curl, CurlWorker &worker)
-{
-    CurlOperation::Setup(curl, worker);
-    curl_easy_setopt(m_curl.get(), CURLOPT_WRITEFUNCTION, CurlCopyOp::WriteCallback);
-    curl_easy_setopt(m_curl.get(), CURLOPT_WRITEDATA, this);
-    curl_easy_setopt(m_curl.get(), CURLOPT_CUSTOMREQUEST, "COPY");
-    m_header_list.reset(curl_slist_append(m_header_list.release(), (std::string("Source: ") + m_source_url).c_str()));
-    curl_easy_setopt(m_curl.get(), CURLOPT_HTTPHEADER, m_header_list.get());
-}
-
-void
-CurlCopyOp::Success()
-{
-    SetDone(false);
-    if (m_handler == nullptr) {return;}
-    auto status = new XrdCl::XRootDStatus();
-    auto obj = new XrdCl::AnyObject();
-    auto handle = m_handler;
-    m_handler = nullptr;
-    handle->HandleResponse(status, obj);
-}
-
-void
-CurlCopyOp::ReleaseHandle()
-{
-    if (m_curl == nullptr) return;
-    curl_easy_setopt(m_curl.get(), CURLOPT_WRITEFUNCTION, nullptr);
-    curl_easy_setopt(m_curl.get(), CURLOPT_WRITEDATA, nullptr);
-    curl_easy_setopt(m_curl.get(), CURLOPT_CUSTOMREQUEST, nullptr);
-    curl_easy_setopt(m_curl.get(), CURLOPT_HTTPHEADER, nullptr);
-    curl_easy_setopt(m_curl.get(), CURLOPT_XFERINFOFUNCTION, nullptr);
-    m_header_list.reset();
-    CurlOperation::ReleaseHandle();
-}
-
-size_t
-CurlCopyOp::WriteCallback(char *buffer, size_t size, size_t nitems, void *this_ptr)
-{
-    auto me = reinterpret_cast<CurlCopyOp*>(this_ptr);
-    std::string_view str_data(buffer, size * nitems);
-    size_t end_line;
-    while ((end_line = str_data.find('\n')) != std::string_view::npos) {
-        auto cur_line = str_data.substr(0, end_line);
-        if (me->m_line_buffer.empty()) {
-            me->HandleLine(cur_line);
-        } else {
-            me->m_line_buffer += cur_line;
-            me->HandleLine(me->m_line_buffer);
-            me->m_line_buffer.clear();
-        }
-        str_data = str_data.substr(end_line + 1);
-    }
-    me->m_line_buffer = str_data;
-
-    return size * nitems;
-}
-
-void
-CurlCopyOp::HandleLine(std::string_view line)
-{
-    if (line == "Perf Marker") {
-        m_bytemark = -1;
-    } else if (line == "End") {
-        if (m_bytemark > -1 && m_callback) {
-            m_callback->Progress(m_bytemark);
-        }
-    } else {
-        auto key_end_pos = line.find(':');
-        if (key_end_pos == line.npos) {
-            return; // All the other callback lines should be of key: value format
-        }
-        auto key = line.substr(0, key_end_pos);
-        auto value = ltrim_view(line.substr(key_end_pos + 1));
-        if (key == "Stripe Bytes Transferred") {
-            try {
-                m_bytemark = std::stoll(std::string(value));
-            } catch (...) {
-                // TODO: Log failure
-            }
-        } else if (key == "success") {
-            m_sent_success = true;
-        } else if (key == "failure") {
-            m_failure = value;
-        }
-    }
-}
-
-CurlPutOp::CurlPutOp(XrdCl::ResponseHandler *handler, const std::string &url, const char *buffer, size_t buffer_size, struct timespec timeout, XrdCl::Log *logger)
-    : CurlOperation(handler, url, timeout, logger),
-    m_data(buffer, buffer_size)
-{
-}
-
-CurlPutOp::CurlPutOp(XrdCl::ResponseHandler *handler, const std::string &url, XrdCl::Buffer &&buffer, struct timespec timeout, XrdCl::Log *logger)
-    : CurlOperation(handler, url, timeout, logger),
-    m_owned_buffer(std::move(buffer)),
-    m_data(buffer.GetBuffer(), buffer.GetSize())
-{
-
-}
-
-void
-CurlPutOp::Setup(CURL *curl, CurlWorker &worker)
-{
-    m_curl_handle = curl;
-    CurlOperation::Setup(curl, worker);
-
-    curl_easy_setopt(m_curl.get(), CURLOPT_UPLOAD, 1);
-    curl_easy_setopt(m_curl.get(), CURLOPT_READDATA, this);
-    curl_easy_setopt(m_curl.get(), CURLOPT_READFUNCTION, CurlPutOp::ReadCallback);
-    if (m_object_size >= 0) {
-        curl_easy_setopt(m_curl.get(), CURLOPT_INFILESIZE_LARGE, m_object_size);
-    }
-}
-
-void
-CurlPutOp::ReleaseHandle()
-{
-    curl_easy_setopt(m_curl.get(), CURLOPT_READFUNCTION, nullptr);
-    curl_easy_setopt(m_curl.get(), CURLOPT_READDATA, nullptr);
-    curl_easy_setopt(m_curl.get(), CURLOPT_UPLOAD, 0);
-    // If one uses just `-1` here -- instead of casting it to `curl_off_t`, then on Linux
-    // we have observed compilers casting the `-1` to an unsigned, resulting in the file
-    // size being set to 4294967295 instead of "unknown".  This causes the second use of the
-    // handle to claim to upload a large file, resulting in the client hanging while waiting
-    // for more input data (which will never come).
-    curl_easy_setopt(m_curl.get(), CURLOPT_INFILESIZE_LARGE, static_cast<curl_off_t>(-1));
-    CurlOperation::ReleaseHandle();
-}
-
-void
-CurlPutOp::Pause()
-{
-    SetDone(false);
-    if (m_handler == nullptr) {
-        m_logger->Warning(kLogXrdClCurl, "Put operation paused with no callback handler");
-        return;
-    }
-    auto handle = m_handler;
-    auto status = new XrdCl::XRootDStatus();
-    m_handler = nullptr;
-    m_owned_buffer.Free();
-    // Note: As soon as this is invoked, another thread may continue and start to manipulate
-    // the CurlPutOp object.  To avoid race conditions, all reads/writes to member data must
-    // be done *before* the callback is invoked.
-    handle->HandleResponse(status, nullptr);
-}
-
-void
-CurlPutOp::Success()
-{
-    SetDone(false);
-    if (m_handler == nullptr) {
-        m_logger->Warning(kLogXrdClCurl, "Put operation succeeded with no callback handler");
-        return;
-    }
-    auto status = new XrdCl::XRootDStatus();
-    auto handle = m_handler;
-    m_handler = nullptr;
-    handle->HandleResponse(status, nullptr);
-}
-
-bool
-CurlPutOp::ContinueHandle()
-{
-    if (!m_curl_handle) {
-		return false;
-	}
-
-	CURLcode rc;
-	if ((rc = curl_easy_pause(m_curl_handle, CURLPAUSE_CONT)) != CURLE_OK) {
-		m_logger->Error(kLogXrdClCurl, "Failed to continue a paused handle: %s", curl_easy_strerror(rc));
-		return false;
-	}
-	return m_curl_handle;
-}
-
-bool
-CurlPutOp::Continue(std::shared_ptr<CurlOperation> op, XrdCl::ResponseHandler *handler, const char *buffer, size_t buffer_size)
-{
-    if (op.get() != this) {
-        Fail(XrdCl::errInternal, 0, "Interface error: must provide shared pointer to self");
-        return false;
-    }
-    m_handler = handler;
-    m_data = std::string_view(buffer, buffer_size);
-    if (!buffer_size)
-    {
-        m_final = true;
-    }
-
-    try {
-        m_continue_queue->Produce(op);
-    } catch (...) {
-        Fail(XrdCl::errInternal, ENOMEM, "Failed to continue the curl operation");
-        return false;
-    }
-    return true;
-}
-
-bool
-CurlPutOp::Continue(std::shared_ptr<CurlOperation> op, XrdCl::ResponseHandler *handler, XrdCl::Buffer &&buffer)
-{
-    if (op.get() != this) {
-        Fail(XrdCl::errInternal, 0, "Interface error: must provide shared pointer to self");
-        return false;
-    }
-    m_handler = handler;
-    m_data = std::string_view(buffer.GetBuffer(), buffer.GetSize());
-    if (!buffer.GetSize())
-    {
-        m_final = true;
-    }
-
-    try {
-        m_continue_queue->Produce(op);
-    } catch (...) {
-        Fail(XrdCl::errInternal, ENOMEM, "Failed to continue the curl operation");
-        return false;
-    }
-    return true;
-}
-
-size_t CurlPutOp::ReadCallback(char *buffer, size_t size, size_t n, void *v) {
-	// The callback gets the void pointer that we set with CURLOPT_READDATA. In
-	// this case, it's a pointer to an HTTPRequest::Payload struct that contains
-	// the data to be sent, along with the offset of the data that has already
-	// been sent.
-	auto op = static_cast<CurlPutOp*>(v);
-    //op->m_logger->Debug(kLogXrdClPelican, "Read callback with buffer %ld and avail data %ld", size*n, op->m_data.size());
-
-    // TODO: Check for timeouts.  If there was one, abort the callback function
-    // and cause the curl worker thread to handle it.
-
-	if (op->m_data.empty()) {
-		if (op->m_final) {
-			return 0;
-		} else {
-			op->Pause();
-			return CURL_READFUNC_PAUSE;
-		}
-	}
-
-	size_t request = size * n;
-	if (request > op->m_data.size()) {
-		request = op->m_data.size();
-	}
-
-	memcpy(buffer, op->m_data.data(), request);
-	op->m_data = op->m_data.substr(request);
-
-	return request;
 }
